Refactor: delete user refactor

+++ esturide_api/contexts/user/domain/service/user_domain_service.py
@@ -15,7 +15,10 @@ class UserService:
         return self.user_repository.get_all_users()
 
     def delete_user(self, user_id: int):
-        return self.user_repository.delete_user(user_id)
+        if self.user_repository.get_user_by_id(user_id):
+            self.user_repository.delete_user(user_id)
+            return {"message": f"User with id {user_id} was successfully deleted"}
+        return {"message": f"User with id {user_id} was not found"}
 
-    def update_user_put(self, user_id: int, updated_user: UserUpdatePut):
+    def update_user_put(self, user_id: int, updated_data: UserUpdatePut):
         pass
